libtimeinstate: check for invalid buckets when reading map
The indexing logic in getUidsUpdatedConcurrentTimes relies on the keys
read from the BPF map being valid and assumes that the maximum bucket
number is determined by the number of CPUs. Add a check to fail
immediately if an invalid, higher bucket is encountered.
Also add a new test in libtimeinstate_test to check this case.

cherry-pick from:https://android-review.googlesource.com/c/platform
/frameworks/native/+/1468306

Change-Id: I431276f5cf928d0663b7fcd751c82bb875eebf63